+++ extensions/shiori-sh/CHANGELOG.md
@@ -4,4 +4,8 @@
 
 - Added AI extension support.
 
+## [0.1.1] - 2026-03-03
+
+- Add action to update link details
+
 ## [Initial Version] - 2026-02-26